refactor: move system sets to their own file
