+++ src/core/sets/mod.rs
@@ -1,0 +1,13 @@
+mod game_over_menu;
+mod game_play;
+mod main_menu;
+mod menu;
+mod pause_menu;
+
+pub use self::{
+    game_over_menu::*,
+    game_play::*,
+    main_menu::*,
+    menu::*,
+    pause_menu::*,
+};
